Correctly handle the case when platform sources are not published.
For Q, sources are not yet published, therefore the platform
installation status should not be shown as ‘Partially installed’
even when the corresponding SDK and all the system images are
installed and up to date.

For API < 14, we already handle that correctly - this CL expands
the scope to handle it for any platform that doesn't have sources
published.

Test: Added testSummaryTreeNodeWithAndWithoutSources
Bug: b/132473200, b/128845709

Change-Id: Icf8df4d49483a3ab532e44a8502882b88ed0427d