Extract lazy subscription, relevance logic from SampleDataListener
SampleDataListener does a lot of work to avoid hogging the EDT in
response to irrelevant VFS events, including not subscribing itself
until it's needed by some caller and using a two-pass system for
checking relevance that allows for "fail fast" behavior based on file
extensions.

This logic is all generally useful for writing well-behaved
VFS listeners for the Android plugin, so this CL extracts these bits
into two new classes: PoliteAndroidVirtualFileListener, a base class that
generalizes the relevance-checking logic, and LazyVirtualFileListenerSubscriber,
a utility for ensuring the listener has been subscribed upon request.

Bug: 132629996
Test: SampleDataResourceRepositoryTest (TODO: test directly)
Change-Id: I0dbad48883e60c173792c05360100a3bc94d5825