[Gradle Build Model] Fix resolving relative paths
This issue was caused by the applied files using the itself as a the
context to resolve any relative paths. Within Gradle everything is run
in the context of the first script since these scripts are just calling
out to other scripts. Via 'apply' statments.

To fix this issue we maintain a stack of parsed files and when
resolving relative paths we use the buttom of the stack for the context.
This ensures that the correct context is used while parsing applied
files.

Bug: 122885755
Test: unit test added
Change-Id: I5e533a0d4fcb4fee82fcde7d313e4e17f260d928