Add process name field to cpu's StopProfilingApp and StopMonitoringApp rpc
The daemon's cpu capture cache has been updated to be keyed by the app's name
instead of pid. Passing down the process name in the associated rpcs to stop
ongoing captures. We cannot rely on the daemon getting the pid from the package
name because the app might have died.

Bug: 119261457
Test: Existing
Change-Id: I1af419ecf1add553b6063d6585736297ddb1b003