Remove checkAppProfilingState and use CpuTraceInfo events as source of truth
- Instead of running a separate Updatable that checks for the profiling state on
every update, CpuProfilerStage now uses the most recently seen CpuTraceInfo to
synchronize the capture state and profilng configuration. This removes a significant
amount of logic/info that can be derived from the CpuTraceInfo messages.
- The logic to handle updating of the parsing state has been moved to be within
CpuCaptureParser. Callers to the parser shouldn't be responsible for telling the
parser what state it is currently in.
- Startup profiling track has been moved to AndroidProfilerLaunchTaskContributor,
at a time where the user performs the action instead of relying on a polling logic
inside the CpuProfilerStage.

Bug: 119261457
Test: Updated
Change-Id: I582e21e01efca4846362b488e89397a50226962b