Do not ignore non-root source-less grouping modules
which are created by Gradle sync to represent module groups and may be
used in JPS projects as well.

Note, that the change to build file nodes in nestedProjects.txt is not
intended, but it is from one broken state to another broken state. The
problem of incorrectly identified root project build file will be
addressed separately.

Bug: 124358808
Test: n/a
Change-Id: Ie922a0eb8a348b68ac5c06c3bc9d365c9bcb7ba9