Include new modules in its own line
We are currently adding new modules to the first include directive we
found in the settings.gradle project file, doing this have some issues
with version control and comments. Also there is an implicit limit of
254 modules that can be pass in a single include.

This CL causes new modules to be included in its own line at the end of
the file. Note that since we cannot parse the files, trying to do more
sophisticated things, like trying to add after the last existing include
would have some corner cases that could cause a failure.

Current implementation also have issues with comments, for example
trying to add a new module to this file would have no effect since it
will add the new module to the commented out include:

/*
include ':commentedOut'
*/

include ':app'

Bug: 133578918
Test: New unit test for RecipeMergeUtils#mergeGradleSettingsFile
Change-Id: I6bcecc51e0d0985690a63edfd9454faee416e4bf