Ignore ProcessCanceledException during execution of Gradle task.
This exception happens if Gradle build is interrupted. There was
a catch block to ignore this exception in the code, but it required
the error message to contain prefix "java.lang.Throwable".
Based on external crash reports, the textual description doesn't
contain the prefix. This change reduce the required prefix to be just
the class name of ProcessCanceledException.

Bug: 122855869
Test: NA
Change-Id: I613bf6ed5a0dc6c72c158ba3138646c23eba2974