Override contentsChanged in PoliteAndroidVirtualFileListener
SampleDataListener is both a PoliteAndroidVirtualFileListener and a
PsiTreeChangeListener, so it doesn't need to respond to VFS content
change events that it will have already picked up as PSI events.
But since PoliteAndroidVirtualFileListener is a base class for VFS
listeners in general, it should respond to these events by default
and leave it to implementing classes to override contentsChanged if
they don't need to respond.

Bug: N/A
Test: SampleDataResourceRepositoryTest still passes
Change-Id: Ife5b728b70144f2e99ba92da4c4371c2f3f91125