Test if ResourceFolderRepository is updated after DumbMode
When the IDE enters or exit dumb mode, the cache mapping the
VirtualFile directories to PsiDirectory FileManagerImpl#getVFileToPsiDirMap
is cleared from the DumbModeListener.

Dumb mode is entered any time a file is added or deleted.

When a file is created in a directory that is not cached in this map,
PsiVFSListener#fileCreated will trigger a PsiTreeChangeEvent.PROP_UNLOADED_PSI
event which is not handled by the ResourceFolderRepository.PsiListener
so the new file is never added to the repository.

Instead of relying on the Psi system for file change event, we now rely on the
VFS system which does not suffer from this problem.

Bug: 130800515
Test: ResourceFolderRepositoryTest.testRepositoryUpdatedAfterDumbMode
Change-Id: I3c4b243c8d02cbea4203a1e363cda3c96ae7c139