Instrumentation test support for dynamic features
In this CL, in order to support instrumentation test for dynamic features:

1) 'Deploy ApkFromBundle' is adopted here. When 'getDeployTasks', Apks from
bundle are explictily collected, instead of single dynamic feature module Apk.

2) ':app:extractApksForDebug' and ':dynamicfeature:assembleDebugAndroidTest'
are the gradle tasks to execute when instrumented test for Dynamic Features. Base-app
module should be retrieved when 'findTasksToExecute'. Also, build action of base-app
module should be created for post build models.

3) When getting application ID for Dynamic Feature module, package name of base-app
application should be retrieved. However, for test package name, package name of
Dynamic Feature module is applied.

4) Testing dynamic features on devices API < 21 is not supported currently.
Error is given.

TODO:

1) As apk from bundle is adopted here, we should make sure users declare properly
in the manifest to guarantee, apk is to be installed. (Maybe warning should be given
instead of method/class not found error message)

2) When using API in base, some dependencies(say, appcompat) might have to be re-added.

3) For original instrumented test for base-app, we do multi-apk deployment(This is legacy way).
However, instead of just instrumented tests belonging to base-app module,
all tests belonging to dynamic feature modules are included in gradle tasks.
Meaning, all tests apks are generated and deployed. I am not sure if this is a proper
behavior or not.

Though './gradlew :app:connectedDebugAndroidTest' only installs
 -- production base Apk: 'app-debug.apk'
 -- testApk: 'app-debug-androidTest.apk'

Bug: 120089420
Test: GradleApkProviderTest.java
      GradleApplicationIdProviderTest.java
      GradleTaskRunnerFactoryTest.java
      AndroidTestRunConfigurationTest.java

Change-Id: If52844c22ff4788e1c4fed2246be7292a401f680