Update the text field on focus gain to avoid overwrite.
When using the resource picker to select a value the text field
used in the editor control in the properties panel may sometimes
gain focus temporarily. This seems to happen when the adding a
new value in "Declared Attributes".

What happens:
1) The selected value from the resource picker is used to add this
   attribute to XML.
2) The text editor (inside the drawable editor) regains focus
   temporarily. The text in the editor is not updated.
3) The text editor loses focus and the editor model decides to
   commit the changes from the editor which is still the empty
   string.

The end result: Nothing changes !

This change is touching some very sensitive editing logic. It would
be great if the reviewer (and QA) would test general editing after
this change.

Fixes: 134434166
Bug: 133184665
Test: unit test added.
Change-Id: I21d51c6c3b98a3559e8b25748dc3df690ec8a559