[PSD] refresh selection on DependenciesReloaded
If we get a DependenciesReloaded event in the
Declared Dependencies Panel, the dependency with which our selection
is associated in the Model will become invalid.  We should therefore
update the selection to refer to its new equivalent, if such a
dependency exists.

In order to do this, fire the DependenciesReloaded event after the
parsed dependency information has been invalidated or recomputed, so
that the consumer of the event can restore their now-invalid state using
the updated dependencies information.  Make this happen at the right
time for PsJavaModules by calling refresh() explicitly.

Bug: 130912720
Test: manual
Change-Id: I2eca154a52b8e44cecf61b1ca70a2202a5362058