Obtain module level dependencies via the module system API
The scope containing the dependencies of a module can be obtained from
the module itself. This CL introduces a level of indirection to this:
rather than obtaining that scope from the module, it obtains it via the
module system API. This allows build systems to customize the set of
dependencies provided via the API (which may differ from the
dependencies expressed to IntelliJ). The Gradle implementation simply
falls back to the existing behavior.

Bug: 111609848
Test: no effective changes for Gradle, see companion CL for others
Change-Id: Ia31ec8cb0fa22e2bc40b73b9f89e39b14faa58f0