Allow 'dp' unit input in SingleWidgetView
Allow a user to type-in a margin value with 'dp' suffix.
The reason is that people might already be used to type-in their
dimensions including the unit, and having it reset to 0 may end up
frustratring to use.

E.g: If they type '24dp' the input should be accepted and write the
value in the attribute, instead of setting it to 0.

Bug: 135262185
Test: manually tested
Change-Id: Iaf45bd74156bf8c2f2c4a838a6d06681de63bfc7