[Dsl] mechanism for proposing property variable names
In preparation for suggesting good variable names for dependencies in
the PSD, implement a mechanism for propagating a reasonable suggestion
from the definition of a property through to where the value might be
extracted as a variable.

It's not totally straightforward because, for example, artifact
versions are such a property, but for the user a good variable name
for the artifact version will include the name of the artifact, and
not simply "version" or anything associated with the kind of value.
Define getPreferredVariableName(), which defaults to returning the
existing default variable name of "var", and propagate it through
ModelProperty objects as appropriate.

Bug: 135231380
Test: existing ExtractVariableWorkerTest
Change-Id: Ia00a796dcdb29984441ef0a75cd9e4a5d8ca3e5c