[Gradle Sync] Don't search in build files straight after sync
We shouldn't need to search in build files as the information should be
present in the model we just fetched from Gradle. This change adds a
method to AndroidPluginInfo to only check the model.

This change then calls this method in AndroidPluginVersionReader which
is only run as a result of the VersionCompatibilityChecker which is
called from PostSyncSetup so, unless sync fails, the model will be
available.

Bug: 135201784
Test: unit test added
Change-Id: Id799e0a57fb823c4557ef51539079dac42f746b4
(cherry picked from commit 687589dc9d2872d8b1e35996720b1e8a9974685b)