Don't block on merged manifest in AndroidImportFilter
AndroidImportFilter#getApplicationPackage blocks on the merged manifest
but can be frequently called from the EDT, leading to UI freezes and
stuttering (usually around 1s, but as bad as over a minute for some
modules).

This CL resolves the issue by having getApplicationPackage() call
AndroidManifestUtils#getPackageName, which uses a cached snapshot of
the merged manifest or manually parses the primary manifest's PSI
if no snapshot is available, rather than using the deprecated
blocking method from MergedManifestManager.

Fixes: 135140820
Test: AndroidImportFilterTest
Change-Id: I23311657ab6eb6f67c04523e1c3369d237a31a55