Strongly reference PsiFiles in MergedManifestInfo
Currently, MergedManifestInfo tracks the modification stamps of the
PsiFiles corresponding to the files that contribute to a merged
manfiest. PsiFiles are weakly referenced from VirtualFiles, however,
so the modification stamps are meaningless for PsiFiles that get garbage
collected, leading to invalidation errors in
MergedManifestRefreshListener.

This CL resolves the issue by strongly referencing the PsiFile instances
as keys of the modification count map in MergedManifestInfo so that the
modification counts remain accurate.

Bug: 135140820
Test: LightClassesTests (once Ib71842993 is submitted)
Change-Id: I86bb2a489fa1146af82edd833eb527005030fb7e