Fix classloader issues in UI tests on JDK 11
For the reasons described in issue 135142609, the defining ClassLoader
of GuiTestLauncher is no longer an instance of UrlClassLoader in JDK 9+
when tests are run from the IDE. Thus, the method previously employed
for extracting the classpath for the client IDE no longer works. In this
case, the java.class.path system property on the server contains all of
the requisite classes.

Also fix finding the java executable for JDK 9+.

Fixes: 135142609
Change-Id: I03b9ad9e66ed7825bb6980773fd70f56016c4a14