Move layeredimage out of uibuilder package
LayeredImageEditor is responsible for handling PSD files, so it's part
of design tools. However, it's not part of uibuilder and needs to be
taken out of this subpackage, which will eventually be moved into its
own separate module. This CL moves layeredimage to another subpackage of
the designer module (designer.editors).

Bug: N/A
Test: N/A
Change-Id: I2aa07e9914dc6e20f7b763ecbfb8376424c733ad