[Psi,.pro] smarter parser
(and less work in the lexer)

The existing functionality is broadly preserved, including handling of
(single- and double-) quoted filename arguments.  The exact error
messages for invalid syntax, and whether they are from the lexer or
the parser, are not preserved; test cases have been updated (and are
expected to be volatile in the face of further lexer and parser
improvements).

Test: updated ProguardParserTest
Bug: 72461769
Change-Id: I198e9739baa27157e2cf6ae683637d816d2c05f7