Replace AndroidModel#getDataBindingMode
AndroidModel currently provides a build system-independent API for
determining the level of data binding support for a given module
(no support, androidx, etc). All implementations of this method
should rely on whether or not the module depends on data binding,
so we can just use the build system-independent dependency management
APIs provided by AndroidModuleSystem instead.

This CL removes AndroidModel#getDataBindingMode and updates
ModuleDataBinding to use DependencyManagementUtil#dependsOn
instead. Since changes to the data binding support level are no longer
broadcasted as updates to the AndroidFacet, ModuleDataBinding now
listens for project syncs to make sure it's up to date with the
project's dependencies.

Bug: 130358739
Test: DataBindingUtilTest
Change-Id: Icb4bab45ae9863c06151ca25aed733bc3a3b7edc