Add Panel as a step element for Assistant
Currently, StatefulButtons are the only way to display content that can
be controlled by Studio; all other content is hardcoded into the
assistant XML configs. However, some dynamic content might not want to
be attached to an action button, or the button layout doesn't work well
with attached text.

This CL adds a new type of StepElement, which is just a JPanel that can
be extended to be kept track of and initialize things when the panel is
created.

Bug: 126602033
Test: N/A
Change-Id: Id4d6df44bae518f4a37f389b0e4dc7be488afded