[PSD] object expression in QuickSearchComboBox
Intellij's EditorComboBox is not quite a drop-in replacement for
JComboBox, despite being a subclass.  In particular, its behaviour
regarding actionPerformed is substantially different, as it does not
consider focus changes or unmapping as triggering an action, which means
that the existing way of maintaining consistency between the editor
state and the selected item does not work.

It is possible to restore that consistency using EditorComboBox's
DocumentListener and the documentChanged method, but only by ensuring
that the selected item state is maintained after (during) every change
to the document.  With the default implementation of setText() and
setSelectedItem(), this would cause infinite recursion (which is
detected by the Document implementation, triggering an exception); we
can break this recursion and have everything work by simply avoiding
setting the control's text when it is already equal to the updated
value.  We can then add a DocumentListener after the combo box is
created, to maintain the state and fire actionPerformed events for us
to consume.

At present, the (few) uses of QuickSearchComboBox do not rely on this
consistency; the uses are in modal dialogs, where we do not have to
worry about the control being hidden or unfocused.

Bug: 127954783
Test: existing
Change-Id: I192b1ef832211b96b717dfbd0e2d38ddcd94ab54