Better UI for resources with no preview and presentation mode
Resources like strings doesn't make much sense to draw a preview for
them, so it'd be nice to have a better representation for those.
Demo in bug, but there's still more to do for it.

The resource explorer is also not very friendly handling scaling,
this should address at least most of the issues with presentation mode.

Bug: 136190135, 135686734
Test: Manually, tested no preview UI enabling String resources
temporarily
Change-Id: I192b9b66150713778737de487623f16ee09e27c1