Fix "SVS enabled but not supported" with merged sync.
In the case that SVS is enabled by flag but not supported
by Gradle, the model provider still creates a VariantGroup model.
And the presence of VariantGroup model is used to idenfity
if it's SVS or not.

This CL doesn't add VariantGroup if Variant model was not found.
And also cleaned up SVS variable from AndroidModuleModel.

Add integration class to test the cases when SVS is enabled but
not supported.

Fixes: 136250944
Test: SingleVariantSyncWithOlderPluginTest
Change-Id: I7cf307d279a77824640bb6917e7782b6daec355b