Generate Java migration class.
Added JavaMigrationClassGenerator which generates a file containing the
Migration class that produces the update between two versions of a
Room database.
The generation is performed by selecting two JSON
files which contain the description of the two versions of the database
and by selecting the GenerateRoomMigration action from the actions menu.

Test: Will add tests for JavaMigrationClassGenerator in a future CL.
Bug: 136048932
Change-Id: I70b014929ca8213e59617548d1ef2df03f1f5475