[PSD] test resolved dependencies post-quickfix focus
Test that focus is (eventually) restored to the dependency which was
previously focused in the Resolved Dependencies Graph/Tree.

This test is somewhat tricky: there's a lot of updating going on, and
some of the test framework is not as bullet-proof as we might
like (specifically, JTreeFixture.clickPath() is not atomic with
respect to changes to the tree, such as adding items, because it
involves scrolling and clicking).  The extra functionality in the
DependenciesTreeFixture is to provide enough inspection of the UI to
judge when it is reasonable to perform operations without risking the
UI changing because of an asynchronous event between UI actions.

Test: new DependenciesTest.testFocusAfterUpdateQuickFix
Bug: 131766339
Change-Id: I579a8fc582b139f4694bc5957b67e79fe3b4bc4e