[New PSD] Show empty mid-hierarchy modules in PSD
Show undeclared modules that appear to be in the middle of gradle path.
e.g. for :libraries:lib1 structure show libraries module even if it is just a container folder.
Since they are not declared, there is no corresponding parsed model, they are uneditable.
Show just empty stub panel for such modules.

Bug: 134442868
Test: existing/added
Change-Id: Ieef545bffd18b9103a25446e47184ee344ba8296