Add cpu trace import support for new data pipeline
Previously, the imported CpuTraceInfo is saved to the database. This change
caches the CpuTraceInfo locally within the CpuProfilerStage to eliminate the
extra rpc. We can guarantee this approach to work because the CpuTraceInfo
should be a singleton within the imported session. This change allows the profiler
to only stream the begin and end session events during import, and the rest of
the logic is handled within the CpuProfilerStage.

Bug: 136292864
Test: Studio-side tests
Change-Id: I43c4a5ed6a7ca50fee5f69ae82c29c125685c890