[Dsl,.kts] resolve long-form string template entries
Resolving extra properties in kotlinscript will naturally involve
handling long-form ${...} string template entries, because of the
extra["prop"] syntax for looking up ext.prop.

Handle finding the injection for the string template entry
by (recursively) finding the injection for the enclosed expression,
except that the GradleReferenceInjection that is the return value must
have the string template entry itself as its psiElement, and not the
expression: otherwise the filter in
GradleReferenceInjection.injectAll() will exclude our injection.

Bug: 130224064
Test: ExtModelTest.resolveVariableInListProperty
Change-Id: Ib6cc7396ba6bd15c027cbe3e2efcf0c5421ce827