Add command handler for starting and stopping legacy cpu traces
- Added ProxyCommandHandler#shouldHandle(...) as a way to tell the proxy
to delegate the command the the device daemon even if the command type is
registered. This allows us to handle the use case where pre-O atraces still
goes through the daemon, while art traces are handled in the proxy layer
- Added LegacyCpuTraceCommandHandler which mirrors the logic in
StudioLegacyCpuTraceProfiler. Instead of trying to perform significant
refactor in the legacy code for better code sharing, it's easier to copy the
logic given the legacy code will be removed in the near future anyway.

Bug: 119261457
Test: Added

Change-Id: I88686e709873b37dd568fd3cc2eb0554b214c991