Fix java debugger not working with AndroidTestOrchestrator
The root cause of this regression is the package name of
AndroidTestOrchestrator is hard coded in the codebase. After
the AndroidX migration, the orchestrator is re-packaged as
androidx so the function stopped working.

This CL makes the implementation of AndroidTestOrchestratorJavaDebuggerTask
more generalized, renames it to ReattachingDebugConnectorTask, which works
with arbitrary orchestrator.

Fixes: 113305007
Test: manual testing (no unit tests written against those classes so far)
Change-Id: I266d20ebc5b4e24603de05a84aa82c5b6da9e3a8