Make merged manifest overrides build system-independent
MergedManifestInfo#getMergedManifest pulls merged manifest overrides
and placeholders directly from the Gradle model, which isn't present
for projects that use other build systems.

This CL resolves the issue by introducing
AndroidModuleSystem#getManifestOverrides as a build system-independent
way to obtain the overrides and placeholders to pass to the manifest
merger when computing the merged manifest for a particular module.
The default implementation of the new API extracts the
overrides that are available from the AndroidModule just as it would
have before. The Gradle-specific implementation matches what
MergedManifestInfo#getMergedManifest was already doing, except that
it doesn't pointlessly initialize a DOM object for the primary manifest
when determining the version name override.

This change also updates the Manifest XML DOM interface to include
addUsesPermission() and getVersionCode() methods. These will be used
in companion CLs to facilitate testing other implementations.

Test: AndroidModuleInfoTest
Fixes: 128928135
Change-Id: If2776b873edbd0f8f17f48129af21f4d38ac61f1