[Dsl,.kts] support apply(from = ...)
The GradleDsl model we have tests specifically for a particular form
of apply: as a GradleDslExpressionMap with name "apply" (see
GradleDslBlockElement.addParsedElement).  This is a straightforward
abstraction from the Groovy DSL, which has an apply application
statement with named arguments, implicitly translated to a single map
argument.  The Kotlin DSL, however, implements apply as a regular
method call with named arguments; in order to make things work,
implement the necessary glue to incorporate the new GradleDsl element
kinds into the apply block.

The test cases also require an implementation of getValue() for
GradleDslLiterals representing the array form of extra properties, and
minor adjustments to test code to account for pathname differences.

Bug: 130224064
Test: GradleDslFile.* translated to kts
Change-Id: I5d62b0a0fc56afe9def1898b7392a6f915a6fe20