Fix AndroidProcessHandler doesn't terminate itself when target processes are finished
As it is documented in AndroidProcessHandler javadoc, this handler should terminate
once all the target remote processes on Android devices are finished if
monitorRemoteProcesses flag is set to true (which is true by default). However, it
does not terminate at all in the current implementation causing several issues such
as instrumentation test with AndroidTestOrchestrator never finishes, "stop" execution
button is displayed even after the android process finishes, etc.

Also, it turned out monitorRemoteProcesses flag was introduced as a workaround solution
for an old bug, b/37083961. AndroidTestRunConfiguration is the only place where it sets
the flag to false. Due to the bug explained above, this flag almost does nothing in
current implementation, and this workaround is inappropriate fix after all. I confirmed
that the original bug b/37083961 doesn't reproduce without this hack anymore so this CL
makes the flag NOOP officially and annotates it as deprecated. I will delete it at all
in my next CL (separate CL because it affects to many files).

This CL splits AndroidProcessHandler into three pieces: SingleDeviceAndroidProcessMonitor,
AndroidProcessMonitorManager and AndroidProcessHandlerImpl. SDAPM is responsible for
monitoring processes running on a device. APMManager manages multiple SDAPM instances.
APHI uses APMManager to support multiple devices concurrently.

Test: Unit tests added
Fixes: 138400182
Change-Id: I92389c9fb7beb0e87d1fc2b253e5add5a36c8922