Fix failing QrScanningCujTest.
QrScanningCujTest uses the play-services-vision library, which uses
AndroidX. As per bug 130286699, the Android Gradle plugin now requires
any projects using AndroidX to declare android.useAndroidX=true in the
gradle.properties file, so this CL adds that property to the test
project.

A more proper fix would be to also upgrade all the dependencies used in
the test project to AndroidX, but that will probably happen at some
point anyway.

Fixes: 138547623
Bug: 130286699
Test: Fixing it
Change-Id: I61b69c61b00034c5918c8cfcb494ca7a5f3eeae7