Support preselecting type in CreateResourceFileAction
Since we are using the CreateResourceFileActions in the ResourceManager,
the intention is no longer to have CreateTypeResourceFileAction depend
on the current directory selected, but to be able to support having a
target ResourceType. And then it should create the directory when
needed.

For typed resource files (eg: 'Create Drawable Resource File') it should
now show the root element and allows to change when possible (vector,
animated-vector, etc...). Should at least help clarify what's going on
and would help the developer not have to manually change it afterwards
if it's not what they wanted.

Fixed Help button in these dialogs.

Bug: 138257967
Test: CreateTypedResourceFileActionTest
Change-Id: I589a89954aa0f3f57d6dc995ec49a100ef3beae2