[Res Explorer] Show OpenTranslationsEditor action
Adding logic to have 'external' actions being displayed on the Resource
Explorer. E.g: Translations Editor, Themes Editor (rip).

Adding just an action for the Translations Editor.
Should show up when on the Strings tab and when selecting a string valid
for the translations editor (seems that it has to be a resource in a
strings.xml file under res/values).

Demo in bug.

Bug: 138755813
Test: n/a UI change for the most part
Change-Id: Icb5af0e108cbbda71363c51b92d62b69de09c44a