[Dsl,.kts] methodCall name lookups in assignment lvalues
Building a GradleNameElement from the left hand side of an assignment
almost works: the Kotlin syntax is close enough to what the
Groovy-inspired GradleNameElement can cope with that all seems well...

... except that idiomatic block lookup in Kotlin involves calling
methods to get blocks by name, rather than simply being able to
dereference a property on its parent.  For example

  android.buildTypes.getByName("foo").applicationIdSuffix = "bar"

where GradleNameElement assumes a dot-qualified identifier with some
array dereferencing suffixes, but explicitly no embedded method calls.

Treat GradleNameElement as an API of sorts, and walk the left-hand
side of assignments to construct the appropriate string that will
suitably initialize GradleNameElement to correspond to the semantics
of the Kotlin source, as far as that is possible.

Bug: 130224064
Test: BuildTypeModelTest, BuildTypesElementTest for kts
Change-Id: I161042a4eadcc7b472435b4774eaa8c3cf10da21