Get JAVA_HOME from terminal for JDK location
We are currently getting it from studio environment and that can be
different to what a user will see in a terminal. We suggest our users to
use the same JDK Location in Studio and in CLI but if we use Studio
environment then that suggestion may be incorrect.

This CL first tries to get the environment variable from a terminal, and
then from current environment.

Test: Covered by IdeSdksAndroidTest.
Bug: 135968300
Bug: 132951215
Change-Id: Ib01664ae624cf37dd6bdd441a91308315f5a5939