Fix AdaptiveIconPreviewTest UI test
AdaptiveIconPreviewTest relies on the fact we use a XML file with the
preview panel, so it should only work when the split editor is not in use.
This CL overrides NELE_SPLIT_EDITOR to false within this test's scope, and
b/138996982 is tracking the addition of new tests to make sure adaptive
icons are supported in the split editor as well.

Bug: 136174865
Bug: 138996982
Test: AdaptiveIconPreviewTest
Change-Id: I6c801f2814a427e619e8bfe0494619678ecf1f82